Move insert query into actions folder
