+++ src/components/main-content/shopping-cart/CartButton/CartButton.tsx
@@ -1,4 +1,5 @@
 "use client";
+import { addProduct } from "@/src/app/lib/actions";
 import CartPreview from "@/src/components/main-content/shopping-cart/CartPreview/CartPreview";
 import { Button, Select, Option } from "@/src/tailwind";
 import { createClient } from "@/src/utils/supabase/supabaseClient";
@@ -47,21 +48,9 @@ export default function CartButton({ singleProduct, cart }: CartButtonProps) {
       setIsFormComplete(false);
       return;
     }
-    const supabase = createClient();
     const { id, title, price, image } = singleProduct;
-    const { error } = await supabase.from("shopping_cart").insert({
-      id,
-      title,
-      price,
-      image,
-      quantity: selectedQuantity,
-      size: selectedSize,
-    });
 
-    if (error) {
-      throw error;
-    }
-
+    addProduct(id, title, price, image, selectedQuantity, selectedSize);
     router.refresh();
     toggleSidebar();
     setSelectedQuantity("");
